pyproject.toml
[tool.poetry]
name = "glpiconnect"
version = "0.1.0"
description = ""
authors = ["Ashley Felton <ashley@ropable.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.0.7"
mysqlclient = "^2.0.1"
django-extensions = "^3.0.8"
dj_database_url = "^0.5.0"
django-dotenv = "^1.4.2"
gunicorn = "^20.0.4"
dj_static = "^0.0.6"

[tool.poetry.dev-dependencies]
ipython = "^7.18.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
